Refactor ActiveAdmin InterventionZone helper to a proper module
And include it properly
